refactor: Add friend button to user profile in dashboard
This commit adds a friend button to the user profile section in the dashboard. The button is displayed next to the user's avatar and allows other users to send a friend request. The functionality is implemented in the addfriend.js file. This change enhances the user experience by providing a convenient way to connect with other users.